requirements.txt
argparse
colorama
cupy-cuda118
deform_conv2d_onnx_exporter
graphviz
ipdb
kornia
lightning
matplotlib
monai
numba
numpy
onnxruntime-gpu
opencv-python
pycuda
pyquaternion
pytorch-lightning[extra]
pyyaml
scipy
sklearn
tqdm
typing-extensions>=4.0.0
wandb